requirements.txt
anthropic>=0.40.0
alpaca-py>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
